Replace make check and distcheck with make test (#2)

+++ utils/utils.h
@@ -12,7 +12,7 @@
 #include "../cBLAS/cBLAS.h"
 
 // Wall time in linux
-double wtime();
+long wtime();
 
 /**
  * Computes norm( A * x - (b - z), 2)
